requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
scipy>=1.7.0

# Geospatial libraries
rasterio>=1.2.10
pyproj>=3.1.0
shapely>=1.8.0
geopandas>=0.10.0
xarray>=0.19.0
rioxarray>=0.9.0
netCDF4>=1.5.7
dask[array]>=2021.9.0
affine>=2.3.0

# AWS integration
boto3>=1.18.0
s3fs>=2021.8.0

# Optional: Database integration
psycopg2-binary>=2.9.1
SQLAlchemy>=1.4.0
GeoAlchemy2>=0.9.0

# Optional: Visualization
folium>=0.12.1
plotly>=5.1.0
holoviews>=1.14.5
geoviews>=1.9.1
cartopy>=0.20.0

# Development tools
jupyter>=1.0.0
jupyterlab>=3.1.0
